impl1/app_ai/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
aiohttp==3.9.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   starlette
attrs==23.2.0
    # via aiohttp
black==24.2.0
    # via -r .\requirements.in
certifi==2023.11.17
    # via
    #   -r .\requirements.in
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via
    #   httpie
    #   requests
click==8.1.7
    # via
    #   black
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   httpie
    #   pytest
    #   tqdm
coverage[toml]==7.4.1
    # via pytest-cov
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
defusedxml==0.7.1
    # via httpie
distro==1.9.0
    # via openai
dnspython==2.5.0
    # via
    #   -r .\requirements.in
    #   pymongo
docopt==0.6.2
    # via -r .\requirements.in
fastapi==0.109.0
    # via -r .\requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.2
    # via httpx
httpie==3.2.2
    # via -r .\requirements.in
httpx==0.26.0
    # via
    #   -r .\requirements.in
    #   openai
hypercorn==0.16.0
    # via -r .\requirements.in
hyperframe==6.0.1
    # via h2
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via rdflib
jinja2==3.1.3
    # via -r .\requirements.in
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.7
    # via -r .\requirements.in
langchain-community==0.0.20
    # via
    #   -r .\requirements.in
    #   langchain
langchain-core==0.1.23
    # via
    #   langchain
    #   langchain-community
langsmith==0.0.87
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.4
    # via jinja2
marshmallow==3.20.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   httpie
    #   yarl
mypy-extensions==1.0.0
    # via
    #   black
    #   typing-inspect
numpy==1.26.4
    # via
    #   -r .\requirements.in
    #   langchain
    #   langchain-community
    #   pandas
openai==1.10.0
    # via -r .\requirements.in
packaging==23.2
    # via
    #   black
    #   langchain-core
    #   marshmallow
    #   pytest
pandas==2.2.0
    # via -r .\requirements.in
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
pluggy==1.4.0
    # via pytest
priority==2.0.0
    # via hypercorn
psutil==5.9.8
    # via -r .\requirements.in
pydantic==2.6.0
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.16.1
    # via pydantic
pygments==2.17.2
    # via
    #   httpie
    #   rich
pymongo==4.6.1
    # via -r .\requirements.in
pyparsing==3.1.1
    # via rdflib
pysocks==1.7.1
    # via requests
pytest==7.3.2
    # via
    #   -r .\requirements.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r .\requirements.in
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.1
    # via -r .\requirements.in
python-multipart==0.0.6
    # via -r .\requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
rdflib==7.0.0
    # via -r .\requirements.in
regex==2023.12.25
    # via tiktoken
requests[socks]==2.31.0
    # via
    #   -r .\requirements.in
    #   httpie
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via httpie
rich==13.7.0
    # via httpie
six==1.16.0
    # via
    #   isodate
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.27
    # via
    #   langchain
    #   langchain-community
starlette==0.35.1
    # via fastapi
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.5.2
    # via -r .\requirements.in
tqdm==4.66.1
    # via openai
typing-extensions==4.9.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2024.1
    # via pandas
urllib3==2.2.0
    # via requests
uvicorn==0.27.0.post1
    # via -r .\requirements.in
wsproto==1.2.0
    # via hypercorn
xmlformatter==0.2.6
    # via -r .\requirements.in
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
